+++ source/app/scorearea.cpp
@@ -73,7 +73,7 @@ void ScoreArea::renderDocument(const Document &document)
 
     myScoreInfoBlock = ScoreInfoRenderer::render(score.getScoreInfo(), activePalette->text().color());
 
-    myRenderedSystems.reserve(score.getSystems().size());
+    myRenderedSystems.reserve(static_cast<int>(score.getSystems().size()));
     for (unsigned int i = 0; i < score.getSystems().size(); ++i)
         myRenderedSystems.append(nullptr);
